+++ app/models/applikation/forms/benefit.rb
@@ -1,0 +1,13 @@
+module Applikation
+  module Forms
+    class Benefit < ::FormObject
+      def self.permitted_attributes
+        { benefits: Boolean }
+      end
+
+      define_attributes
+
+      validates :benefits, inclusion: { in: [true, false] }
+    end
+  end
+end
