+++ spec/models/applikation/forms/benefit_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Forms::Benefit do
+RSpec.describe Applikation::Forms::Benefit do
   params_list = %i[benefits]
 
   let(:hash) { {} }
