+++ spec/models/form_object_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
-RSpec.describe Forms::Base do
+RSpec.describe FormObject do
 
   # rubocop:disable ClassAndModuleChildren
-  class Forms::TestClass < Forms::Base
+  class FormTestClass < FormObject
     def self.permitted_attributes
       { id: Integer, fee: Integer }
     end
@@ -11,11 +11,11 @@ RSpec.describe Forms::Base do
     define_attributes
   end
 
-  params_list = Forms::TestClass.permitted_attributes.keys
+  params_list = FormTestClass.permitted_attributes.keys
 
   describe 'when Application object is passed in' do
     let(:application) { create :application }
-    let(:form) { Forms::TestClass.new(application) }
+    let(:form) { FormTestClass.new(application) }
 
     params_list.each do |attr_name|
       it "assigns #{attr_name}" do
@@ -26,7 +26,7 @@ RSpec.describe Forms::Base do
 
   describe 'when a Hash is passed in' do
     let(:hash) { { id: 1, fee: 2 } }
-    let(:form) { Forms::TestClass.new(hash) }
+    let(:form) { FormTestClass.new(hash) }
 
     params_list.each do |attr_name|
       it "assigns #{attr_name}" do
